requirements.txt
# Paquetes principales
fastapi==0.118.0
uvicorn==0.37.0
pydantic==2.11.9
python-multipart==0.0.20

# LangChain y Ollama
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.78
langchain-ollama==0.3.10
langchain-text-splitters==0.3.11
langchain-together==0.3.0
langsmith==0.4.32

# Procesamiento de documentos
PyPDF2==3.0.1
pypdf==6.1.1

# Vectorización y embeddings
sentence-transformers==5.1.1
faiss-cpu==1.12.0
huggingface-hub==0.35.3

# Utilidades y dependencias complementarias
aiohttp==3.12.15
async-timeout==4.0.3
requests==2.32.5
typing-extensions==4.15.0
numpy==2.2.6
scikit-learn==1.7.2
scipy==1.15.3
SQLAlchemy==2.0.43
tenacity==9.1.2
httpx==0.28.1
httpcore==1.0.9
requests-toolbelt==1.0.0
python-dotenv==1.1.1
dataclasses-json==0.6.7
pydantic-settings==2.11.0
attrs==25.3.0
colorama==0.4.6
tqdm==4.67.1
filelock==3.19.1
regex==2025.9.18
fsspec==2025.9.0
greenlet==3.2.4
idna==3.10
packaging==25.0
PyYAML==6.0.3
sympy==1.14.0
torch==2.8.0
transformers==4.57.0
tokenizers==0.22.1
typing-inspect==0.9.0
typing-inspection==0.4.2
sniffio==1.3.1
aiohappyeyeballs==2.6.1
yarl==1.20.1
zstandard==0.25.0

# Testing
pytest
pytest-asyncio
httpx
pytest-cov